+++ src/components/collections/Header/utils/getDesktopMenuMarkup.spec.js
@@ -1,14 +1,5 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { Menu } from 'semantic-ui-react';
-import {
-  expectComponentToBe,
-  expectComponentToHaveChildren,
-  expectComponentToHaveProps,
-} from '@lodgify/enzyme-jest-expect-helpers';
-
-import { Submenu } from 'elements/Submenu';
-import { Button } from 'elements/Button';
 
 import { navigationItems } from '../mock-data/navigationItems';
 
@@ -18,7 +9,7 @@ const getMarkupAsRenderedComponent = extraProps =>
   shallow(
     <div>
       {getDesktopMenuMarkup({
-        activeNavigationItemIndex: 0,
+        activeNavigationItemIndex: 1,
         navigationItems,
         ...extraProps,
       })}
@@ -26,88 +17,21 @@ const getMarkupAsRenderedComponent = extraProps =>
   );
 
 describe('getDesktopMenuMarkup', () => {
-  it('should render the right children', () => {
-    const wrapper = getMarkupAsRenderedComponent();
+  describe('by default', () => {
+    it('should render the right structure', () => {
+      const actual = getMarkupAsRenderedComponent();
 
-    expectComponentToHaveChildren(wrapper, Menu.Item, Submenu);
-  });
-
-  describe('the `Submenu` component', () => {
-    const getSubmenu = extraProps =>
-      getMarkupAsRenderedComponent(extraProps).find(Submenu);
-
-    it('should get the right props', () => {
-      const { subItems, text } = navigationItems[1];
-      const wrapper = getSubmenu();
-
-      expectComponentToHaveProps(wrapper, {
-        isMenuItem: true,
-        isSimple: true,
-        isTriggerUnderlined: false,
-        isTriggeredOnHover: true,
-        items: subItems,
-        children: text,
-      });
-    });
-
-    it('should get `props.isTriggerUnderlined = true` if required', () => {
-      const wrapper = getSubmenu({
-        activeNavigationItemIndex: 1,
-      });
-
-      expectComponentToHaveProps(wrapper, { isTriggerUnderlined: true });
+      expect(actual).toMatchSnapshot();
     });
   });
 
   describe('if `props.primaryCTA` is passed', () => {
     const primaryCTA = { onClick: Function.prototype, text: 'someText' };
-    const getMenuItem = () =>
-      getMarkupAsRenderedComponent({ primaryCTA })
-        .find(Menu.Item)
-        .at(1);
 
-    it('should render an extra Semantic UI `Menu.Item`', () => {
-      const wrapper = getMarkupAsRenderedComponent({ primaryCTA })
-        .children()
-        .at(2);
+    it('should render the right structure', () => {
+      const actual = getMarkupAsRenderedComponent({ primaryCTA });
 
-      expectComponentToBe(wrapper, Menu.Item);
-    });
-
-    describe('the `Menu.Item` component ', () => {
-      it('should get the right props', () => {
-        const wrapper = getMenuItem();
-
-        expectComponentToHaveProps(wrapper, {
-          className: 'no-underline',
-          link: true,
-        });
-      });
-
-      it('should render a Lodgify UI `Button` component', () => {
-        const wrapper = getMenuItem();
-
-        expectComponentToHaveChildren(wrapper, Button);
-      });
-
-      describe('the Lodgify UI `Button`', () => {
-        const getButton = () => getMenuItem().find(Button);
-
-        it('should have the right props', () => {
-          const wrapper = getButton();
-
-          expectComponentToHaveProps(wrapper, {
-            isRounded: true,
-            onClick: expect.any(Function),
-          });
-        });
-
-        it('should have the right children', () => {
-          const wrapper = getButton();
-
-          expectComponentToHaveChildren(wrapper, primaryCTA.text);
-        });
-      });
+      expect(actual).toMatchSnapshot();
     });
   });
 });