+++ src/components/collections/Header/utils/getMobileMenuMarkup.spec.js
@@ -1,15 +1,5 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { Menu, Accordion } from 'semantic-ui-react';
-import {
-  expectComponentToBe,
-  expectComponentToHaveChildren,
-  expectComponentToHaveProps,
-} from '@lodgify/enzyme-jest-expect-helpers';
-
-import { SearchBar } from 'general-widgets/SearchBar';
-import { Modal } from 'elements/Modal';
-import { Icon, ICON_NAMES } from 'elements/Icon';
 
 import { navigationItems } from '../mock-data/navigationItems';
 
@@ -44,77 +34,11 @@ const getMarkupAsRenderedComponent = extraProps =>
     .first();
 
 describe('getMobileMenuMarkup', () => {
-  it('should render a Semantic UI `Menu.Item`', () => {
-    const wrapper = getMarkupAsRenderedComponent();
+  describe('by default', () => {
+    it('should render the right structure', () => {
+      const actual = getMarkupAsRenderedComponent();
 
-    expectComponentToBe(wrapper, Menu.Item);
-  });
-
-  describe('the `Menu.Item` component', () => {
-    it('should render a Lodgify UI `Modal` component', () => {
-      const wrapper = getMarkupAsRenderedComponent();
-
-      expectComponentToHaveChildren(wrapper, Modal);
-    });
-  });
-
-  describe('the `Modal` component', () => {
-    const getModal = () => getMarkupAsRenderedComponent().find(Modal);
-
-    it('should have the right props', () => {
-      const wrapper = getModal();
-
-      expectComponentToHaveProps(wrapper, {
-        isFullscreen: true,
-        trigger: <Icon name={ICON_NAMES.BARS} />,
-      });
-    });
-
-    it('should have the right children', () => {
-      const wrapper = getModal();
-
-      expectComponentToHaveChildren(wrapper, Menu);
-    });
-  });
-
-  describe('the `Menu` component', () => {
-    const getMenu = () => getMarkupAsRenderedComponent().find(Menu);
-
-    it('should have the right props', () => {
-      const wrapper = getMenu();
-
-      expectComponentToHaveProps(wrapper, {
-        text: true,
-        vertical: true,
-      });
-    });
-
-    it('should have the right children', () => {
-      const wrapper = getMenu();
-
-      expectComponentToHaveChildren(wrapper, Menu.Item, Menu.Item, Accordion);
-    });
-  });
-
-  describe('each `Accordion` component', () => {
-    it('should have the right props', () => {
-      const wrapper = getMarkupAsRenderedComponent().find(Accordion);
-
-      expectComponentToHaveProps(wrapper, {
-        as: Menu.Item,
-        panels: [
-          {
-            title: {
-              content: navigationItems[1].text,
-              key: expect.any(String),
-            },
-            content: {
-              content: expect.any(Array),
-              key: expect.any(String),
-            },
-          },
-        ],
-      });
+      expect(actual).toMatchSnapshot();
     });
   });
 
@@ -125,34 +49,10 @@ describe('getMobileMenuMarkup', () => {
         searchBarLocationOptions,
       });
 
-    it('should render a Semantic UI `Menu.Item` component as the first child of the fragment', () => {
-      const wrapper = getSearchBarMenuItem();
+    it('should render the right structure', () => {
+      const actual = getSearchBarMenuItem();
 
-      expectComponentToBe(wrapper, Menu.Item);
-    });
-
-    describe('the `Menu.Item` component', () => {
-      it('should render a Lodgify UI `SearchBar` component', () => {
-        const wrapper = getSearchBarMenuItem();
-
-        expectComponentToHaveChildren(wrapper, SearchBar);
-      });
-    });
-
-    describe('the `SearchBar` component', () => {
-      it('should have the right props', () => {
-        const wrapper = getSearchBarMenuItem().find(SearchBar);
-
-        expectComponentToHaveProps(wrapper, {
-          guestsOptions: searchBarGuestsOptions,
-          isDisplayedAsModal: true,
-          locationOptions: searchBarLocationOptions,
-          getIsDayBlocked: searchBarGetIsDayBlocked,
-          modalHeadingText: searchBarModalHeadingText,
-          onSubmit: expect.any(Function),
-          searchButton: searchBarSearchButton,
-        });
-      });
+      expect(actual).toMatchSnapshot();
     });
   });
 });