+++ src/components/elements/Submenu/utils/getValueProperty.js
@@ -1,9 +1,0 @@
-/**
- * @param {boolean} isSelectedDisabled
- */
-export const getValueProperty = isSelectedDisabled =>
-  isSelectedDisabled
-    ? {
-        value: -999,
-      }
-    : {};